refactor: move AppProfile testing to test_app_profile.py
